Rename Math to MathHelpers. Fix docker run test command #52
