+++ CDBTo3DTiles/src/CDBElevation.cpp
@@ -1,7 +1,7 @@
 #include "CDBElevation.h"
 #include "BoundingRegion.h"
 #include "Ellipsoid.h"
-#include "Math.h"
+#include "MathHelpers.h"
 #include "glm/gtc/type_ptr.hpp"
 #include "meshoptimizer.h"
