+++ Core/src/GlobeRectangle.cpp
@@ -1,5 +1,5 @@
 #include "GlobeRectangle.h"
-#include "Math.h"
+#include "MathHelpers.h"
 
 namespace Core {
