+++ Core/src/IntersectionTests.cpp
@@ -1,5 +1,5 @@
 #include "IntersectionTests.h"
-#include "Math.h"
+#include "MathHelpers.h"
 
 namespace Core {
