+++ Core/src/MathHelpers.cpp
@@ -1,4 +1,4 @@
-#include "Math.h"
+#include "MathHelpers.h"
 
 namespace Core {
 const double Math::EPSILON1 = 1e-1;
