+++ Core/src/Transforms.cpp
@@ -1,5 +1,5 @@
 #include "Transforms.h"
-#include "Math.h"
+#include "MathHelpers.h"
 
 namespace Core {
 glm::dmat4x4 Transforms::eastNorthUpToFixedFrame(const glm::dvec3 &origin, const Ellipsoid &ellipsoid)
